upgrade library and fix some
