--- core/src/main/scala/com.ariskk.raft/model/ComandResponse.scala
@@ -1,6 +1,0 @@
-package com.ariskk.raft.model
-
-sealed trait CommandResponse
-case object Committed                 extends CommandResponse
-case class Redirect(leaderId: NodeId) extends CommandResponse
-case object LeaderNotFoundResponse    extends CommandResponse
