--- core/src/main/scala/com.ariskk.raft/model/Command.scala
@@ -1,5 +1,0 @@
-package com.ariskk.raft.model
-
-sealed trait Command extends Serializable
-trait ReadCommand    extends Command
-trait WriteCommand   extends Command
