+++ core/src/main/scala/com/ariskk/raft/model/Command.scala
@@ -1,0 +1,8 @@
+package com.ariskk.raft.model
+
+sealed trait Command extends Serializable
+object Command {
+  trait ReadCommand extends Command
+
+  trait WriteCommand extends Command
+}
