+++ core/src/main/scala/com/ariskk/raft/model/CommandResponse.scala
@@ -1,0 +1,8 @@
+package com.ariskk.raft.model
+
+sealed trait CommandResponse
+object CommandResponse {
+  case object Committed                 extends CommandResponse
+  case class Redirect(leaderId: NodeId) extends CommandResponse
+  case object LeaderNotFoundResponse    extends CommandResponse
+}
