+++ core/src/main/scala/com/ariskk/raft/model/ElectionTimeout.scala
