+++ core/src/main/scala/com/ariskk/raft/model/Index.scala
