+++ core/src/main/scala/com/ariskk/raft/model/Message.scala
