+++ core/src/main/scala/com/ariskk/raft/model/NodeId.scala
