+++ core/src/main/scala/com/ariskk/raft/model/NodeState.scala
@@ -1,0 +1,13 @@
+package com.ariskk.raft.model
+
+sealed trait NodeState
+
+object NodeState {
+
+  case object Follower  extends NodeState
+  case object Candidate extends NodeState
+  case object Leader    extends NodeState
+
+  lazy val default: NodeState = Follower
+
+}
