+++ core/src/main/scala/com/ariskk/raft/model/Vote.scala
