+++ core/src/main/scala/com/ariskk/raft/statemachine/KeyValueStore.scala
@@ -1,0 +1,17 @@
+package com.ariskk.raft.statemachine
+
+import com.ariskk.raft.model.Command.{ ReadCommand, WriteCommand }
+import zio._
+
+final case class Key(value: String)              extends AnyVal
+final case class ReadKey(key: Key)               extends ReadCommand
+final case class WriteKey[T](key: Key, value: T) extends WriteCommand
+
+final class KeyValueStore[T](map: Ref[Map[Key, T]]) extends StateMachine[T] {
+  def write = { case WriteKey(key: Key, value: T) => map.update(_ + (key -> value)) }
+  def read  = { case ReadKey(key) => map.get.map(_.get(key)) }
+}
+
+object KeyValueStore {
+  def apply[R]: ZIO[Any, Nothing, KeyValueStore[R]] = Ref.make(Map.empty[Key, R]).map(new KeyValueStore(_))
+}
